# perf: decrease allocs