--- a/caller/internal/caller/validators.go
+++ b/caller/internal/caller/validators.go
@@ -26,9 +26,9 @@ import (
 )
 
 var (
-	DontValidate     = ChainValidator{}
-	ValidateWither   = ChainValidator{validateWither}
-	ValidateProvider = ChainValidator{validateProvider}
+	DontValidate     FuncValidator = nil
+	ValidateWither                 = ChainValidator{validateWither}
+	ValidateProvider               = ChainValidator{validateProvider}
 )
 
 var (