**before**

```
BenchmarkContainer_scopeDefault-4                      	 6346165	       184.7 ns/op	     112 B/op	       4 allocs/op
BenchmarkContainer_scopeShared-4                       	 6631849	       179.8 ns/op	     112 B/op	       4 allocs/op
BenchmarkContainer_scopeContextual-4                   	  395386	      3094 ns/op	     896 B/op	      30 allocs/op
BenchmarkContainer_scopeContextualSetField-4           	  685257	      1489 ns/op	     624 B/op	      18 allocs/op
BenchmarkContainer_scopeContextual_in_same_context-4   	10513569	       111.5 ns/op	      16 B/op	       1 allocs/op
BenchmarkContainer_scopeNonShared-4                    	  376460	      2668 ns/op	     608 B/op	      29 allocs/op
BenchmarkContainer_map-4                               	296526640	         4.033 ns/op	       0 B/op	       0 allocs/op
```

**after**

```
BenchmarkContainer_scopeDefault-4                      	 6336008	       188.9 ns/op	     120 B/op	       4 allocs/op
BenchmarkContainer_scopeShared-4                       	 6434148	       189.8 ns/op	     120 B/op	       4 allocs/op
BenchmarkContainer_scopeContextual-4                   	  526902	      2114 ns/op	     760 B/op	      20 allocs/op
BenchmarkContainer_scopeContextualSetField-4           	  758175	      1355 ns/op	     584 B/op	      14 allocs/op
BenchmarkContainer_scopeContextual_in_same_context-4   	10349127	       114.0 ns/op	      24 B/op	       1 allocs/op
BenchmarkContainer_scopeNonShared-4                    	  572878	      1936 ns/op	     472 B/op	      19 allocs/op
BenchmarkContainer_map-4                               	283429525	         4.028 ns/op	       0 B/op	       0 allocs/op
```